Rename file to match cop name
