+++ lib/rubocop/cop/sorbet/binding_constant_without_type_alias.rb
