+++ lib/rubocop/cop/sorbet_cops.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-require_relative "sorbet/binding_constants_without_type_alias"
+require_relative "sorbet/binding_constant_without_type_alias"
 require_relative "sorbet/constants_from_strings"
 require_relative "sorbet/forbid_superclass_const_literal"
 require_relative "sorbet/forbid_include_const_literal"
